+++ Editor/Editor/EditorLayer.h
@@ -6,6 +6,10 @@
 
 #include <Core/EventSystem/Layer.h>
 
+namespace GLEngine::Physics::Primitives {
+struct S_AABB;
+}
+
 namespace GLEngine::Renderer {
 class I_Renderer;
 class I_DebugDraw;
@@ -22,19 +26,21 @@ class I_Input;
 // hold gizmos manager, know about selected 3D entities
 // decide which model will render its AABB box etc.
 namespace GLEngine::Editor {
+class C_AABBEditor;
 class C_CurveEditor;
 class C_AnimationEditor;
 
 class EDITOR_API_EXPORT C_EditorLayer : public Core::C_Layer {
 public:
 	C_EditorLayer(Renderer::I_DebugDraw& dd, const Core::I_Input& directInput, const Renderer::C_Viewport& viewport);
-	~C_EditorLayer();
-	void		 SetCamera(std::shared_ptr<Renderer::I_CameraComponent> camera);
-	virtual void OnEvent(Core::I_Event& event) override;
-	virtual void OnUpdate() override;
-	void		 SetViewport(const Renderer::C_Viewport& viewport);
+	~C_EditorLayer() override;
+	void SetCamera(const std::shared_ptr<Renderer::I_CameraComponent>& camera);
+	void OnEvent(Core::I_Event& event) override;
+	void OnUpdate() override;
+	void SetViewport(const Renderer::C_Viewport& viewport);
 
 	void EditCurve(Renderer::C_Curve& curve);
+	void EditAABB(Physics::Primitives::S_AABB& aabb);
 
 	void EditAnimation(); // TODO!
 
@@ -45,6 +51,7 @@ private:
 	Renderer::C_Viewport					   m_Viewport;
 
 	std::unique_ptr<C_CurveEditor>	   m_CurveEditing;
+	std::unique_ptr<C_AABBEditor>	   m_AABBEditing;
 	std::unique_ptr<C_AnimationEditor> m_AnimationEditing;
 };
 } // namespace GLEngine::Editor