+++ Renderer/Renderer/Colours.h
@@ -16,4 +16,11 @@ inline constexpr static auto cyan	= T_Colour(0, 1, 1);
 inline constexpr static auto yellow = T_Colour(1, 1, 0);
 inline constexpr static auto gray	= T_Colour(169.0 / 255.0, 169.0 / 255.0, 169.0 / 255.0);
 
+namespace Editing {
+
+inline constexpr static auto mouseOver = yellow;
+inline constexpr static auto selected  = blue;
+inline constexpr static auto disabled  = gray;
+} // namespace Editing
+
 } // namespace GLEngine::Colours